pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "lyapunov-lab"
version = "0.1.0"
description = "Chaos Analyzer hardware dashboard"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pyserial",
    "click",
    "python-multipart>=0.0.20",
    "bump2version>=1.0.1",
]

[project.scripts]
lyapunov-lab = "lyapunov_lab.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["lyapunov_lab*"]

[tool.uv.workspace]
members = ["lyapunov_lab/backend"]
